Fix the type of variables that store the return value of roundToZero().
git-svn-id: svn://svn.berlios.de/gale/trunk@1376 2d798ef8-6e06-0410-941a-9092aa770104

+++ gale2/include/gale/math/color.h
@@ -122,7 +122,7 @@ struct ColorModel
         float R,G,B;
 
         if (S>0) {
-            long long i=roundToZero(H);
+            int i=roundToZero(H);
             float f=H-i;
 
             float p=V*(1-S);
